+++ src/entropy/mod.rs
@@ -5,9 +5,11 @@ mod hartley;
 mod min_entropy;
 mod renyi;
 mod shannon;
+mod tsallis;
 
 pub use collision::{collision_entropy, collision_entropy_distribution};
 pub use hartley::{hartley_entropy, hartley_entropy_distribution};
 pub use min_entropy::{min_entropy, min_entropy_distribution};
 pub use renyi::{InvalidRenyiOrder, renyi_entropy, renyi_entropy_distribution};
 pub use shannon::{shannon, shannon_distribution};
+pub use tsallis::{InvalidTsallisOrder, tsallis_entropy, tsallis_entropy_distribution};